requirements.txt
py-cord
audioop-lts
aiohttp
asyncio
python-dotenv